adds sane() to http services
